Feature/refactor app logic (#490)
* modularize app logic

* add management service

* add management service

* add management service
